# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ServerBee API",
    "description": "ServerBee VPS monitoring probe API. All responses are wrapped in `{\"data\": <value>}`. Errors return `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notification-groups": {
      "get": {
        "tags": ["notification-groups"],
        "operationId": "list_notification_groups",
        "responses": {
          "200": {
            "description": "List all notification groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationGroup"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": ["notification-groups"],
        "operationId": "create_notification_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notification group created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/notification-groups/{id}": {
      "get": {
        "tags": ["notification-groups"],
        "operationId": "get_group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationGroup"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": ["notification-groups"],
        "operationId": "update_notification_group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notification group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationGroup"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": ["notification-groups"],
        "operationId": "delete_notification_group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification group deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/notifications": {
      "get": {
        "tags": ["notifications"],
        "operationId": "list_notifications",
        "responses": {
          "200": {
            "description": "List all notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": ["notifications"],
        "operationId": "create_notification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notification created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/notifications/{id}": {
      "get": {
        "tags": ["notifications"],
        "operationId": "get_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": ["notifications"],
        "operationId": "update_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notification updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": ["notifications"],
        "operationId": "delete_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/notifications/{id}/test": {
      "post": {
        "tags": ["notifications"],
        "operationId": "test_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test notification sent"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateNotification": {
        "type": "object",
        "required": ["name", "notify_type", "config_json"],
        "properties": {
          "config_json": {},
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "notify_type": {
            "type": "string"
          }
        }
      },
      "CreateNotificationGroup": {
        "type": "object",
        "required": ["name", "notification_ids"],
        "properties": {
          "name": {
            "type": "string"
          },
          "notification_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "required": [
          "id",
          "name",
          "notify_type",
          "config_json",
          "enabled",
          "created_at"
        ],
        "properties": {
          "config_json": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notify_type": {
            "type": "string"
          }
        }
      },
      "NotificationGroup": {
        "type": "object",
        "required": ["id", "name", "notification_ids_json", "created_at"],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notification_ids_json": {
            "type": "string"
          }
        }
      },
...
      "UpdateNotification": {
        "type": "object",
        "properties": {
          "config_json": {},
          "enabled": {
            "type": ["boolean", "null"]
          },
          "name": {
            "type": ["string", "null"]
          },
          "notify_type": {
            "type": ["string", "null"]
          }
        }
      },
      "UpdateNotificationGroup": {
        "type": "object",
        "properties": {
          "name": {
            "type": ["string", "null"]
          },
          "notification_ids": {
            "type": ["array", "null"],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "api_key": []
    },
    {
      "bearer_token": []
    }
  ],
...
}